Adding all code in main for testing
