--- PongGraphics/include/Game_Draw.h
@@ -1,9 +1,0 @@
-#pragma once
-#include "Game.h"
-#include "Draw_Text.h"
-
-extern float ratio;
-
-void Draw_Ball(T_Ball bll);
-void Draw_Racket(T_Racket racket);
-void Draw_Score(T_Score score);
